\name{eqt.pos}
\alias{eqt.pos}
\title{Calculate the positions of equations in a plot region}
\usage{
eqt.pos(no = 50,col.n = 3)
}
\description{
The function could be used as calcuating the positions of equations in a plot region. 
}
\arguments{
	\item{no}{the number of equations will be shown onto the plot region.}
	\item{col.n}{the columns of equations will be shown, which the order is from top to down and from left to right.}
}
\author{ Guofang Liu \email{liugf@ibcas.ac.cn}
}

\examples{

eqt.pos(no=50, column=4)
}
