\name{score24}
\alias{score24}
\title{Calculate solutions on score-24 game for Chinese kids.}
\usage{
score24(x)
}
\description{
The function could be used as calcuating solutions on score-24 game for Chinese kids.
}
\arguments{
	\item{x}{an array with four elements.}
}
\author{ Guofang Liu \email{liugf@ibcas.ac.cn}
}

\examples{

score24(x = c(1,2,9,11))
}
